created separate reducer for reviews
