+++ frontend/src/store/index.js
@@ -3,11 +3,13 @@ import thunk from "redux-thunk";
 import sessionReducer from './session';
 import vehiclesReducer from "./vehicle";
 import bookingsReducer from "./booking";
+import reviewReducer from "./review";
 
 const rootReducer = combineReducers({
   session: sessionReducer,
   vehicles: vehiclesReducer,
-  bookings: bookingsReducer
+  bookings: bookingsReducer,
+  reviews: reviewReducer
 });
 
 let enhancer;
