+++ frontend/src/store/vehicle.js
@@ -2,7 +2,7 @@ import { csrfFetch } from "./csrf";
 
 const LOAD = 'vehicles/LOAD';
 const LOAD_ONE = 'vehicles/LOAD_ONE'
-const SET_REVIEW = 'vehicles/SET_REVIEW'
+
 
 //action creator
 const loadVehicles = (vehicles) => ({
@@ -15,10 +15,7 @@ const loadOne = (vehicle) => ({
     vehicle
 })
 
-const setReview = (vehicle) => ({
-    type: SET_REVIEW,
-    vehicle
-})
+
 
 export const getVehicles = () => async (dispatch) => {
     const res = await csrfFetch(`/api/vehicles`);
@@ -39,18 +36,7 @@ export const getOneVehicle = (id) => async (dispatch) => {
     }
 }
 
-export const reviewVehicle = (id, payload) => async (dispatch) => {
-    const res = await csrfFetch(`/api/vehicles/${id}`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload)
-    });
 
-    if (res.ok) {
-        const newReview = await res.json();
-        dispatch(setReview(newReview))
-    }
-}
 
 const vehiclesReducer = (state = {}, action) => {
     switch (action.type) {
@@ -62,14 +48,6 @@ const vehiclesReducer = (state = {}, action) => {
                 allVehicles[vehicle.id] = vehicle;
             });
             return allVehicles;
-        }
-
-        case SET_REVIEW: {
-                const newState = {
-                    ...state,
-                    [action?.vehicles?.id]: action.vehicles
-                }
-                return newState
         }
         default:
             return state;
